Improve precision of surface3d worldPosition calculation
When the objectOffset is very large compared to the values in the
localCoordinate, adding it to the localCoordinate loses precision in the
localCoordinate. As the addition of the objectOffset was immediately
undone by application of the model matrix, it is better to combine the
model matrix and the translation by the objectOffset into one matrix and
apply it directly to the localCoordinate.